apps/ai/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.30.6
python-multipart==0.0.9
openai-whisper==20231117
torch==2.3.1
numpy==1.26.4
pyannote.audio==3.3.0
huggingface_hub==0.19.4
